requirements.txt
numpy
opencv-python
mediapipe
torchvision
torchvision
scikit-learn
matplotlib
tqdm